backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
python-dotenv==1.0.1
pydantic>=2.11.7,<3
supabase==2.24.0
email-validator==2.2.0
